# cmd/server/main.go
// vpublish - Go语言Gin框架实现的低空智能平台软件包管理系统
//
// 本API文档描述了vpublish系统的各个接口信息，包含管理员API和APP端API两大类
//
//	Schemes: http, https
//	Host: localhost:8080
//	BasePath: /api/v1
//	Version: 2.0.0
//	Contact: {name: vpublish团队, email: support@example.com}
//	License: MIT {url: http://opensource.org/licenses/MIT}
//
//	SecurityDefinitions:
//	BearerAuth:
//	  type: apiKey
//	  name: Authorization
//	  in: header
//	  description: 仅用于管理员端 API, 格式: Bearer <JWT-token>
//	SignatureAuth:
//	  type: apiKey
//	  name: X-App-Key
//	  in: header
//	  description: 用于APP端 API 的X-App-Key值
//	TimestampAuth:
//	  type: apiKey
//	  name: X-Timestamp
//	  in: header
//	  description: 用于APP端 API 的Unix时间戳
//	SignatureValueAuth:
//	  type: apiKey
//	  name: X-Signature
//	  in: header
//	  description: 用于APP端 API 的HMAC-SHA256签名
//
//	Security:
//	BearerAuth: []
//	SignatureAuth: []
//	TimestampAuth: []
//	SignatureValueAuth: []
//
// swagger:info
package main

import (
	"context"
	"fmt"
	"log"
	"net/http"
	"os"
	"os/signal"
	"syscall"
	"time"

	"github.com/gin-gonic/gin"
	swaggerFiles "github.com/swaggo/files"
	ginSwagger "github.com/swaggo/gin-swagger"
	_ "github.com/taerc/vpublish/docs" // swagger docs (generated by swag init)
	"github.com/taerc/vpublish/internal/config"
	"github.com/taerc/vpublish/internal/database"
	"github.com/taerc/vpublish/internal/handler"
	"github.com/taerc/vpublish/internal/middleware"
	"github.com/taerc/vpublish/internal/repository"
	"github.com/taerc/vpublish/internal/service"
	"github.com/taerc/vpublish/internal/version"
	"github.com/taerc/vpublish/pkg/jwt"
	"github.com/taerc/vpublish/pkg/storage"
)

func main() {
	// 打印版本信息
	log.Printf("Starting vpublish %s", version.String())

	// 加载配置
	configPath := config.ResolveConfigPath("./configs/config.yaml")
	cfg, err := config.Load(configPath)
	if err != nil {
		log.Fatalf("load config: %v", err)
	}

	// 初始化数据库
	db, err := database.New(&cfg.Database)
	if err != nil {
		log.Fatalf("connect database: %v", err)
	}

	// 自动迁移
	if err := database.Migrate(db); err != nil {
		log.Fatalf("migrate database: %v", err)
	}

	// 初始化存储
	store, err := storage.NewLocalStorage(cfg.Storage.Path)
	if err != nil {
		log.Fatalf("init storage: %v", err)
	}

	// 初始化 JWT
	jwtService := jwt.New(cfg.JWT.Secret, cfg.JWT.Expire, cfg.JWT.RefreshExpire)

	// 初始化 Repository
	userRepo := repository.NewUserRepository(db)
	appKeyRepo := repository.NewAppKeyRepository(db)
	mcpCredRepo := repository.NewMCPCredentialRepository(db)
	categoryRepo := repository.NewCategoryRepository(db)
	packageRepo := repository.NewPackageRepository(db)
	versionRepo := repository.NewVersionRepository(db)
	statsRepo := repository.NewStatsRepository(db)
	errorRecordRepo := repository.NewErrorRecordRepository(db)

	// 初始化 Service
	userService := service.NewUserService(userRepo)
	categoryService := service.NewCategoryService(categoryRepo)
	packageService := service.NewPackageService(packageRepo, versionRepo, categoryRepo, store, "", cfg.Storage.ExternalDownloadURLPrefix)
	statsService := service.NewStatsService(statsRepo)
	appKeyService := service.NewAppKeyService(appKeyRepo)
	mcpCredService := service.NewMCPCredentialService(mcpCredRepo)
	errorReportService := service.NewErrorReportService(errorRecordRepo)

	// 初始化 Handler
	authHandler := handler.NewAuthHandler(userService, jwtService)
	userHandler := handler.NewUserHandler(userService)
	categoryHandler := handler.NewCategoryHandler(categoryService)
	packageHandler := handler.NewPackageHandler(packageService, statsRepo, appKeyRepo)
	statsHandler := handler.NewStatsHandler(statsService)
	appKeyHandler := handler.NewAppKeyHandler(appKeyService)
	mcpCredHandler := handler.NewMCPCredentialHandler(mcpCredService)
	errorReportHandler := handler.NewErrorReportHandler(errorReportService)

	// 创建路由
	gin.SetMode(cfg.Server.Mode)
	r := gin.New()
	r.Use(middleware.Logger())
	r.Use(middleware.Recovery())
	r.Use(middleware.CORS(&cfg.CORS))

	setupRoutes(r, authHandler, userHandler, categoryHandler, packageHandler, statsHandler,
		appKeyHandler, mcpCredHandler, errorReportHandler, jwtService, appKeyRepo, cfg.Server.Mode)

	// 启动服务器
	srv := &http.Server{
		Addr:         fmt.Sprintf(":%d", cfg.Server.Port),
		Handler:      r,
		ReadTimeout:  cfg.Server.ReadTimeout,
		WriteTimeout: cfg.Server.WriteTimeout,
	}

	// 优雅关闭
	go func() {
		if err := srv.ListenAndServe(); err != nil && err != http.ErrServerClosed {
			log.Fatalf("listen: %s\n", err)
		}
	}()

	log.Printf("Server started on port %d", cfg.Server.Port)

	// 等待中断信号
	quit := make(chan os.Signal, 1)
	signal.Notify(quit, syscall.SIGINT, syscall.SIGTERM)
	<-quit

	log.Println("Shutting down server...")

	ctx, cancel := context.WithTimeout(context.Background(), 5*time.Second)
	defer cancel()

	if err := srv.Shutdown(ctx); err != nil {
		log.Fatal("Server forced to shutdown:", err)
	}

	log.Println("Server exited")
}

func setupRoutes(
	r *gin.Engine,
	authHandler *handler.AuthHandler,
	userHandler *handler.UserHandler,
	categoryHandler *handler.CategoryHandler,
	packageHandler *handler.PackageHandler,
	statsHandler *handler.StatsHandler,
	appKeyHandler *handler.AppKeyHandler,
	mcpCredHandler *handler.MCPCredentialHandler,
	errorReportHandler *handler.ErrorReportHandler,
	jwtService *jwt.JWT,
	appKeyRepo *repository.AppKeyRepository,
	serverMode string,
) {
	// 健康检查 - 包含版本信息
	r.GET("/health", func(c *gin.Context) {
		c.JSON(200, gin.H{
			"status":  "ok",
			"version": version.String(),
			"info":    version.Get(),
		})
	})

	// 版本信息接口
	r.GET("/version", func(c *gin.Context) {
		c.JSON(200, version.Get())
	})

	// Swagger 文档路由（仅在非生产环境启用）
	if serverMode != gin.ReleaseMode {
		r.GET("/swagger/*any", ginSwagger.WrapHandler(swaggerFiles.Handler))
	}

	// API v1
	v1 := r.Group("/api/v1")
	{
		// ============ 管理端 API ============
		admin := v1.Group("/admin")
		{
			// 认证
			admin.POST("/auth/login", authHandler.Login)
			admin.POST("/auth/refresh", authHandler.RefreshToken)
			admin.POST("/auth/logout", authHandler.Logout)

			// 需要登录的路由
			auth := admin.Group("")
			auth.Use(middleware.JWTAuth(jwtService))
			{
				// 用户信息
				auth.GET("/auth/profile", authHandler.GetProfile)
				auth.PUT("/auth/password", authHandler.ChangePassword)

				// 用户管理
				auth.GET("/users", userHandler.List)
				auth.GET("/users/:id", userHandler.Get)
				auth.POST("/users", userHandler.Create)
				auth.PUT("/users/:id", userHandler.Update)
				auth.DELETE("/users/:id", userHandler.Delete)
				auth.PUT("/users/:id/password", userHandler.ResetPassword)

				// 类别管理
				auth.GET("/categories", categoryHandler.List)
				auth.GET("/categories/active", categoryHandler.ListActive) // 获取启用的类别列表
				auth.GET("/categories/:id", categoryHandler.Get)
				auth.POST("/categories", categoryHandler.Create)
				// auth.PUT("/categories/:id", categoryHandler.Update)    // 暂时禁用编辑功能
				// auth.DELETE("/categories/:id", categoryHandler.Delete) // 暂时禁用删除功能

				// 软件包管理
				auth.GET("/packages", packageHandler.List)
				auth.GET("/packages/:id", packageHandler.Get)
				auth.POST("/packages", packageHandler.Create)
				auth.PUT("/packages/:id", packageHandler.Update)
				auth.DELETE("/packages/:id", packageHandler.Delete)

				// 版本管理
				auth.GET("/packages/:id/versions", packageHandler.ListVersions)
				auth.POST("/packages/:id/versions", packageHandler.UploadVersion)
				auth.DELETE("/versions/:id", packageHandler.DeleteVersion)
				auth.GET("/versions/:id/download", packageHandler.DownloadVersion) // 管理端下载

				// 统计
				auth.GET("/stats/daily", statsHandler.DailyStats)
				auth.GET("/stats/trend", statsHandler.DailyTrend)
				auth.GET("/stats/monthly", statsHandler.MonthlyStats)
				auth.GET("/stats/yearly", statsHandler.YearlyStats)
				auth.GET("/stats/category", statsHandler.CategoryStats)
				auth.GET("/stats/overview", statsHandler.Overview)

				// AppKey 管理
				auth.GET("/appkeys", appKeyHandler.List)
				auth.GET("/appkeys/:id", appKeyHandler.Get)
				auth.POST("/appkeys", appKeyHandler.Create)
				auth.PUT("/appkeys/:id", appKeyHandler.Update)
				auth.DELETE("/appkeys/:id", appKeyHandler.Delete)
				auth.POST("/appkeys/:id/regenerate", appKeyHandler.RegenerateSecret)

				// MCP 凭证管理
				auth.GET("/mcp-credentials", mcpCredHandler.List)
				auth.GET("/mcp-credentials/:id", mcpCredHandler.Get)
				auth.POST("/mcp-credentials", mcpCredHandler.Create)
				auth.PUT("/mcp-credentials/:id", mcpCredHandler.Update)
				auth.DELETE("/mcp-credentials/:id", mcpCredHandler.Delete)
				auth.POST("/mcp-credentials/:id/regenerate", mcpCredHandler.RegenerateSecret)

				// 报错记录管理（需要JWT认证）
				auth.GET("/error/records", errorReportHandler.List)
				auth.GET("/error/records/:id", errorReportHandler.Get)
				auth.PUT("/error/records/:id/remark", errorReportHandler.UpdateRemark)
				auth.GET("/error/modules", errorReportHandler.GetModules)

				// 报错统计（需要JWT认证）
				auth.GET("/error/statistics/trend", errorReportHandler.GetTrend)
				auth.GET("/error/statistics/module", errorReportHandler.GetModuleStats)
			}

			// 报错上报（不需要JWT认证，支持直接上报）
			// APP端建议使用 /api/v1/app/error/report（签名认证）
			// 管理端可以使用 /api/v1/admin/error/report（无需认证）
			admin.POST("/error/report", errorReportHandler.Report)
			admin.POST("/error/report/batch", errorReportHandler.BatchReport)
		}

		// ============ APP端 API ============
		app := v1.Group("/app")
		app.Use(middleware.SignatureAuth(appKeyRepo))
		{
			// 类别列表
			app.GET("/categories", categoryHandler.ListActive)

			// 获取某类别的最新版本
			app.GET("/categories/:code/latest", packageHandler.GetLatestByCategory)

			// 获取某类别的版本列表（最多5个）
			app.GET("/categories/:code/versions", packageHandler.GetVersionsByCategory)

			// 下载
			app.GET("/download/:id", packageHandler.Download)

			// 报错上报
			app.POST("/error/report", errorReportHandler.Report)
			app.POST("/error/report/batch", errorReportHandler.BatchReport)
		}
	}
}