# docs/swagger.yaml
definitions:
...
  handler.UpdateRemarkRequest:
    properties:
      remark:
        description: 备注信息
        example: 已处理，问题已解决
        type: string
    required:
    - remark
    type: object
...
  model.AppKey:
    properties:
      app_key:
        description: AppKey
        example: abc123def456
        type: string
      app_name:
        description: 应用名称
        example: 移动客户端
        type: string
      created_at:
        description: 创建时间
        example: "2024-03-12T10:00:00Z"
        type: string
      description:
        description: 描述
        example: 用于移动APP的API访问密钥
        type: string
      id:
        description: 主键ID
        example: 1
        type: integer
      is_active:
        description: 是否启用
        example: true
        type: boolean
      updated_at:
        description: 更新时间
        example: "2024-03-12T15:30:00Z"
        type: string
    type: object
...
  response.PageData:
    properties:
      list: {}
      page:
        type: integer
      page_size:
        type: integer
      total:
        type: integer
    type: object
  response.Response:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
  service.AppKeyResponse:
    properties:
      app_key:
        type: string
      app_name:
        type: string
      app_secret:
        type: string
      created_at:
        type: string
      description:
        type: string
      id:
        type: integer
      is_active:
        type: boolean
      updated_at:
        type: string
    type: object
...
  service.CreateAppKeyRequest:
    properties:
      app_name:
        maxLength: 100
        minLength: 1
        type: string
      description:
        maxLength: 200
        type: string
    required:
    - app_name
    type: object
...
info:
  contact: {}
paths:
  /admin/appkeys:
    get:
      consumes:
      - application/json
      description: 获取AppKey分页列表，用于管理APP端认证密钥
      parameters:
      - default: 1
        description: 页码
        in: query
        minimum: 1
        name: page
        type: integer
      - default: 20
        description: 每页数量
        in: query
        maximum: 100
        minimum: 1
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功，返回AppKey分页列表
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  allOf:
                  - $ref: '#/definitions/response.PageData'
                  - properties:
                      list:
                        items:
                          $ref: '#/definitions/model.AppKey'
                        type: array
                    type: object
              type: object
        "401":
          description: 未认证访问
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - 'BearerAuth ':
        - ""
      summary: 获取AppKey列表
      tags:
      - 管理员/AppKey管理
    post:
      consumes:
      - application/json
      description: 创建新的APP认证密钥，系统自动生成AppKey和AppSecret。注意：AppSecret仅在创建时返回一次，请妥善保存。
      parameters:
      - description: 创建AppKey请求参数
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/service.CreateAppKeyRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 创建成功，返回AppKey信息（包含AppSecret）
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/service.AppKeyResponse'
              type: object
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: 未认证访问
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - 'BearerAuth ':
        - ""
      summary: 创建AppKey
      tags:
      - 管理员/AppKey管理
  /admin/appkeys/{id}:
    delete:
      consumes:
      - application/json
      description: 删除指定的AppKey，删除后该密钥将无法使用
      parameters:
      - description: AppKey ID
        in: path
        minimum: 1
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 删除成功
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: 请求参数错误，无效的ID
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: 未认证访问
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: AppKey不存在
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - 'BearerAuth ':
        - ""
      summary: 删除AppKey
      tags:
      - 管理员/AppKey管理
    get:
      consumes:
      - application/json
      description: 根据ID获取单个AppKey的详细信息
      parameters:
      - description: AppKey ID
        in: path
        minimum: 1
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功，返回AppKey详情
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/model.AppKey'
              type: object
        "400":
          description: 请求参数错误，无效的ID
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: 未认证访问
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: AppKey不存在
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - 'BearerAuth ':
        - ""
      summary: 获取AppKey详情
      tags:
      - 管理员/AppKey管理
    put:
      consumes:
      - application/json
      description: 更新AppKey信息，可修改应用名称、描述和启用状态
      parameters:
      - description: AppKey ID
        in: path
        minimum: 1
        name: id
        required: true
        type: integer
      - description: 更新AppKey请求参数
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/service.UpdateAppKeyRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 更新成功，返回更新后的AppKey信息
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/model.AppKey'
              type: object
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: 未认证访问
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: AppKey不存在
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - 'BearerAuth ':
        - ""
      summary: 更新AppKey
      tags:
      - 管理员/AppKey管理
  /admin/appkeys/{id}/regenerate:
    post:
      consumes:
      - application/json
      description: 为指定AppKey重新生成AppSecret，旧的Secret将立即失效。注意：新的AppSecret仅在此次返回，请妥善保存。
      parameters:
      - description: AppKey ID
        in: path
        minimum: 1
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 重新生成成功，返回新的AppSecret
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/service.AppKeyResponse'
              type: object
        "400":
          description: 请求参数错误，无效的ID
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: 未认证访问
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: AppKey不存在
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - 'BearerAuth ':
        - ""
      summary: 重新生成AppSecret
      tags:
      - 管理员/AppKey管理
  /admin/auth/login:
    post:
      consumes:
      - application/json
      description: 用户通过用户名和密码完成身份认证，成功后返回JWT token和用户信息
      parameters:
      - description: 登录请求参数
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 登录成功，返回token和用户信息
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/handler.LoginResponse'
              type: object
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: 认证失败，用户名或密码错误
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/response.Response'
      summary: 用户登录
      tags:
      - 管理员/认证
...
  /admin/auth/password:
    put:
      consumes:
      - application/json
      description: 用户修改自己的登录密码，需要提供旧密码和新密码。新密码长度要求8-72个字符。
      parameters:
      - description: 修改密码请求参数
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/service.ChangePasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 密码修改成功
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: 请求参数错误或旧密码不正确
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: 未认证，需要有效的JWT token
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - 'BearerAuth ':
        - ""
      summary: 修改当前用户密码
      tags:
      - 管理员/认证
  /admin/auth/profile:
    get:
      consumes:
      - application/json
      description: 获取当前已登录用户的详细信息，包括用户名、昵称、邮箱、角色等
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功，返回用户信息
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/handler.ProfileResponse'
              type: object
        "401":
          description: 未认证，需要有效的JWT token
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: 用户不存在
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - 'BearerAuth ':
        - ""
      summary: 获取当前登录用户信息
      tags:
      - 管理员/认证
  /admin/auth/refresh:
    post:
      consumes:
      - application/json
      description: 使用刷新令牌获取新的访问令牌，刷新令牌通过请求头X-Refresh-Token传递
      parameters:
      - description: 刷新令牌
        in: header
        name: X-Refresh-Token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 刷新成功，返回新的访问令牌
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  additionalProperties: true
                  type: object
              type: object
        "401":
          description: 刷新令牌无效或已过期
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/response.Response'
      summary: 刷新访问令牌
      tags:
      - 管理员/认证
  /admin/categories:
    get:
      consumes:
      - application/json
      description: 获取软件类别的分页列表，支持按页码和每页数量进行分页查询
      parameters:
      - default: 1
        description: 页码
        in: query
        minimum: 1
        name: page
        type: integer
      - default: 50
        description: 每页数量
        in: query
        maximum: 100
        minimum: 1
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功，返回类别分页列表
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  allOf:
                  - $ref: '#/definitions/response.PageData'
                  - properties:
                      list:
                        items:
                          $ref: '#/definitions/model.Category'
                        type: array
                    type: object
              type: object
        "401":
          description: 未认证访问
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - 'BearerAuth ':
        - ""
      summary: 获取类别分页列表
      tags:
      - 管理员/类别管理
    post:
      consumes:
      - application/json
      description: 创建一个新的软件类别，系统会自动根据中文名称生成拼音代码（如：无人机 -> TYPE_WU_REN_JI）
      parameters:
      - description: 创建类别的请求参数
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.CreateCategoryRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 创建成功，返回新创建的类别信息
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/model.Category'
              type: object
        "400":
          description: 请求参数错误，例如参数未满足约束条件或类别名称已存在
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: 未认证，需要有效的管理员Token
          schema:
            $ref: '#/definitions/response.Response'
        "403":
          description: 权限不足，当前用户角色不允许此操作
          schema:
            $ref: '#/definitions/response.Response'
        "409":
          description: 冲突，相同名称的类别已存在
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - 'BearerAuth ':
        - ""
      summary: 创建软件类别
      tags:
      - 管理员/类别管理
  /admin/categories/{id}:
    delete:
      consumes:
      - application/json
      description: 删除指定ID的软件类别，软删除操作
      parameters:
      - description: 类别ID
        example: 1
        in: path
        minimum: 1
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 删除成功
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: 请求参数错误，无效的类别ID
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: 未认证，需要有效的管理员Token
          schema:
            $ref: '#/definitions/response.Response'
        "403":
          description: 权限不足，当前用户角色不允许此操作
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: 类别不存在
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - 'BearerAuth ':
        - ""
      summary: 删除软件类别
      tags:
      - 管理员/类别管理
    get:
      consumes:
      - application/json
      description: 通过ID获取指定软件类别的详细信息
      parameters:
      - description: 类别ID
        example: 1
        in: path
        minimum: 1
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 返回类别详细信息
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/model.Category'
              type: object
        "400":
          description: 请求参数错误，无效的类别ID
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: 未认证访问
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: 类别不存在
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - 'BearerAuth ':
        - ""
      summary: 根据ID获取类别详情
      tags:
      - 管理员/类别管理
    put:
      consumes:
      - application/json
      description: 更新指定ID的软件类别信息，如果更新名称则系统会自动重新生成拼音代码
      parameters:
      - description: 类别ID
        example: 1
        in: path
        minimum: 1
        name: id
        required: true
        type: integer
      - description: 更新类别的请求参数
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.UpdateCategoryRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 更新成功，返回更新后的类别信息
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/model.Category'
              type: object
        "400":
          description: 请求参数错误，无效的类别ID或参数不满足约束条件
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: 未认证，需要有效的管理员Token
          schema:
            $ref: '#/definitions/response.Response'
        "403":
          description: 权限不足，当前用户角色不允许此操作
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: 类别不存在
          schema:
            $ref: '#/definitions/response.Response'
        "409":
          description: 冲突，相同名称的类别已存在
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - 'BearerAuth ':
        - ""
      summary: 更新软件类别
      tags:
      - 管理员/类别管理
...
  /admin/error/modules:
    get:
      consumes:
      - application/json
      description: 获取系统中所有的报错模块列表，用于筛选条件
      produces:
      - application/json
      responses:
        "200":
          description: 返回模块列表
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  items:
                    type: string
                  type: array
              type: object
        "401":
          description: 未认证访问
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - 'BearerAuth ':
        - ""
      summary: 获取所有报错模块列表
      tags:
      - 管理员/报错管理
  /admin/error/records:
    get:
      consumes:
      - application/json
      description: 分页查询报错记录列表，支持按时间、应用类型、模块、报错类型等条件筛选
      parameters:
      - default: 1
        description: 页码
        in: query
        minimum: 1
        name: page
        type: integer
      - default: 20
        description: 每页数量
        in: query
        maximum: 100
        minimum: 1
        name: page_size
        type: integer
      - description: 开始时间戳（毫秒）
        in: query
        name: start_time
        type: integer
      - description: 结束时间戳（毫秒）
        in: query
        name: end_time
        type: integer
      - description: 应用类型：app/platform
        in: query
        name: app_type
        type: string
      - description: 报错模块
        in: query
        name: module
        type: string
      - description: 报错类型：system_error/business_error
        in: query
        name: error_type
        type: string
      - description: 关键词搜索
        in: query
        name: keyword
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功，返回报错记录分页列表
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  allOf:
                  - $ref: '#/definitions/response.PageData'
                  - properties:
                      list:
                        items:
                          $ref: '#/definitions/model.ErrorRecord'
                        type: array
                    type: object
              type: object
        "401":
          description: 未认证访问
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - 'BearerAuth ':
        - ""
      summary: 获取报错记录分页列表
      tags:
      - 管理员/报错管理
  /admin/error/records/{id}:
    get:
      consumes:
      - application/json
      description: 根据ID获取报错记录的详细信息
      parameters:
      - description: 报错记录ID
        example: 1
        in: path
        minimum: 1
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 返回报错记录详情
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/model.ErrorRecord'
              type: object
        "400":
          description: 请求参数错误，无效的ID
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: 未认证访问
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: 报错记录不存在
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - 'BearerAuth ':
        - ""
      summary: 获取报错记录详情
      tags:
      - 管理员/报错管理
  /admin/error/records/{id}/remark:
    put:
      consumes:
      - application/json
      description: 更新指定报错记录的备注信息
      parameters:
      - description: 报错记录ID
        example: 1
        in: path
        minimum: 1
        name: id
        required: true
        type: integer
      - description: 更新备注请求参数
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.UpdateRemarkRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 更新成功
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: 未认证访问
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: 报错记录不存在
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - 'BearerAuth ':
        - ""
      summary: 更新报错记录备注
      tags:
      - 管理员/报错管理
  /admin/error/report:
    post:
      consumes:
      - application/json
      description: 上报单条接口报错信息到平台，系统会自动识别报错类型（无需JWT认证）
      parameters:
      - description: 报错上报请求参数
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.ErrorReportRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 上报成功，返回创建的报错记录
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/model.ErrorRecord'
              type: object
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/response.Response'
      summary: 单条报错上报
      tags:
      - 管理员/报错管理
...
  /admin/stats/overview:
    get:
      consumes:
      - application/json
      description: 获取系统下载统计的概览信息，包含今日下载量、本月下载量、本年下载量和按类别的下载分布
      produces:
      - application/json
      responses:
        "200":
          description: 返回统计概览，包含today、monthly、yearly和by_category字段
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  additionalProperties: true
                  type: object
              type: object
        "401":
          description: 未授权访问
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - 'BearerAuth ':
        - ""
      summary: 获取统计概览
      tags:
      - 管理员/统计
...
swagger: "2.0"